requirements.txt
keras
tensorflow
Pillow
gunicorn
